Refactoring of package contents and class names (#105)
